added requirements.txt. created models and apps. recreated database
Follow steps in requirements.txt before doing anything. Use the sql file in the github folder not the one in the email.
Do NOT run migrations. I'm not done tweaking the models yet. did not run inspectdb either, plan to run inspectdb on a test project and check if it generates the proper models.

Next Task after migrations:
1) Test if Django changes reflect on database